requirements.txt
Pillow>=10.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.10.0
tiktoken>=0.5.2
